TLS SDK 版本更新
* TLS Java支持检查必填参数
* Java SDK 支持重试策略优化
* 支持易用版PutLogs接口
* remove duplicate validation
* 添加PutLogsV2的Demo和Producer接口
* 优化报错提示
* 搜索、直方图、下载任务接口request参数bigint改为long
* Merge remote-tracking branch 'origin/master' into real_2.3.1
* 优化重试策略
* bugfix

https://code.byted.org/volcengine/volc-sdk-java/merge_requests/495

+++ volc-sdk-java/src/test/java/com/volcengine/service/tls/impl/LogsTest.java
@@ -51,8 +51,8 @@ public class LogsTest extends BaseTest {
                     {
                         request.setQuery("*");
                         request.setTopicId(topicId);
-                        request.setStartTime(BigInteger.valueOf(System.currentTimeMillis() - 1000 * 10));
-                        request.setEndTime(BigInteger.valueOf(System.currentTimeMillis()));
+                        request.setStartTime(System.currentTimeMillis() - 1000 * 10);
+                        request.setEndTime(System.currentTimeMillis());
                         request.setInterval(BigInteger.valueOf(1000));
                     }
                     client.describeHistogram(request);
@@ -150,8 +150,8 @@ public class LogsTest extends BaseTest {
             searchLogsRequest.setTopicId(topicId);
             searchLogsRequest.setQuery("test");
             //开始时间20220701
-            searchLogsRequest.setStartTime(BigInteger.valueOf(1656604800000L));
-            searchLogsRequest.setEndTime(BigInteger.valueOf(System.currentTimeMillis()));
+            searchLogsRequest.setStartTime(1656604800000L);
+            searchLogsRequest.setEndTime(System.currentTimeMillis());
             SearchLogsResponse searchLogsResponse = client.searchLogs(searchLogsRequest);
             assertTrue(searchLogsResponse.getLogs().size() > 0);
 
@@ -171,8 +171,8 @@ public class LogsTest extends BaseTest {
                 {
                     request.setQuery("*");
                     request.setTopicId(topicId);
-                    request.setStartTime(BigInteger.valueOf(System.currentTimeMillis() - 1000 * 10));
-                    request.setEndTime(BigInteger.valueOf(System.currentTimeMillis()));
+                    request.setStartTime(System.currentTimeMillis() - 1000 * 10);
+                    request.setEndTime(System.currentTimeMillis());
                     request.setInterval(BigInteger.valueOf(1000));
                 }
                 DescribeHistogramResponse response = client.describeHistogram(request);
